poker-agent-api/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
python-jose[cryptography]==3.3.0
requests==2.32.3
redis==5.2.0
python-socketio[client]==5.11.0
